requirements.txt
imap-tools==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
schedule==1.2.1
six==1.16.0
